fix a bit
